+++ tests/test_donut_viz_pipeline_cwfs_sensors.py
@@ -17,7 +17,9 @@ from lsst.donut.viz import (
     PlotDonutCwfsTaskConfig,
     PlotPsfZernTask,
     PlotPsfZernTaskConfig,
+    PlotDonutFitsTask,
 )
+from lsst.donut.viz.utilities import get_day_obs_seq_num_from_visitid
 from lsst.obs.lsst import LsstCam
 from lsst.ts.wep.task import DonutStamps
 from lsst.ts.wep.utils import (
@@ -96,7 +98,7 @@ class TestDonutVizPipeline(TestCase):
         agg_zern_avg = self.butler.get(average_dataset_list[0])
         self.assertEqual(len(agg_zern_avg), 1)
         self.assertCountEqual(agg_zern_avg["detector"], self.extraDetectorNames)
-        self.assertCountEqual(agg_zern_avg.meta.keys(), self.meta_keys)
+        self.assertCountEqual(agg_zern_avg.meta.keys(), self.meta_keys + ["estimatorInfo"])
 
     def testAggregateZernikesRaw(self):
         raw_dataset_list = list(
@@ -187,7 +189,7 @@ class TestDonutVizPipeline(TestCase):
         self.assertEqual(len(avg_visit_table_list), 1)
         self.assertEqual(avg_visit_table_list[0].dataId["visit"], 4021123106000)
         avg_visit_table = self.butler.get(avg_visit_table_list[0])
-        self.assertCountEqual(avg_visit_table.meta.keys(), self.meta_keys)
+        self.assertCountEqual(avg_visit_table.meta.keys(), self.meta_keys + ["estimatorInfo"])
         avg_zern_table = self.butler.get(
             "aggregateZernikesAvg",
             dataId=avg_visit_table_list[0].dataId,
@@ -259,6 +261,52 @@ class TestDonutVizPipeline(TestCase):
         dataset_list = list(self.butler.query_datasets("donutFits", collections=self.test_run_name))
         self.assertEqual(len(dataset_list), 1)
         self.assertEqual(dataset_list[0].dataId["visit"], 4021123106000)
+
+    def testPlotDonutFitsTaskRunMissingMetadata(self):
+        table_ref = list(
+            self.butler.registry.queryDatasets(
+                "aggregateAOSVisitTableRaw", collections=self.test_run_name
+            ).expanded()
+        )[0]
+        table = self.butler.get(table_ref)
+        stamps_extra = self.butler.get(
+            self.butler.query_datasets("donutStampsExtraVisit", collections=self.test_run_name)[0]
+        )
+        stamps_intra = self.butler.get(
+            self.butler.query_datasets("donutStampsIntraVisit", collections=self.test_run_name)[0]
+        )
+        camera = LsstCam().getCamera()
+
+        self.task = PlotDonutFitsTask()
+        day_obs, seq_num = get_day_obs_seq_num_from_visitid(4021123106000)
+        record = table_ref.dataId.records["visit"]
+
+        table.meta["estimatorInfo"] = {"fwhm": [1.5, 1.5]}
+        # Test that median shifts are output to log
+        with self.assertLogs(logger=self.task.log.logger, level="WARNING") as cm:
+            self.task.run(table, stamps_intra, stamps_extra, camera, day_obs, seq_num, record)
+        for idx, warn_output in zip(range(2), cm.output):
+            err_msg = str(
+                "WARNING:lsst.plotDonutFitsTask:No model plot produced for R00, "
+                + f"donut index: {idx}. Required metadata for danish model not found in "
+                + "aggregateAOSVisitTableRaw."
+            )
+            self.assertEqual(warn_output, err_msg)
+
+        # Test getModel function
+        err_msg = str(
+            "danish_meta must contain the following keys: "
+            + "['fwhm', 'model_dx', 'model_dy', 'model_sky_level'], but only contains: {'fwhm'}"
+        )
+        with self.assertRaises(ValueError) as cm:
+            self.task.getModel(
+                table[0]["zk_CCS"],
+                table.meta["nollIndices"],
+                table.meta["estimatorInfo"],
+                stamps_extra,
+                stamps_intra,
+            )
+        self.assertEqual(str(cm.exception), err_msg)
 
     def testPlotDonutCwfsTask(self):
         # Test that plots exist in butler